created separated component for canvas view
