+++ iDibujo/iDibujo/Model/Line.swift
@@ -1,0 +1,15 @@
+//
+//  Line.swift
+//  iDibujo
+//
+//  Created by Israel Manzo on 6/16/24.
+//  Copyright © 2024 Israel Manzo. All rights reserved.
+//
+
+import SwiftUI
+
+struct Line {
+    var points: [CGPoint]
+    var color: Color
+}
+
