+++ iDibujo/iDibujo/View/MainCanvas.swift
@@ -8,11 +8,6 @@
 
 import SwiftUI
 
-struct Line {
-    var points: [CGPoint]
-    var color: Color
-}
-
 struct MainCanvas: View {
     
     @State private var lines = [Line]()
@@ -27,28 +22,7 @@ struct MainCanvas: View {
                 }
                 clearButton()
             }
-            Canvas { context, size in
-                lines.forEach {
-                    var path = Path()
-                    path.addLines($0.points)
-                    
-                    let strokeStyle = StrokeStyle(lineWidth: 5, lineCap: .round, lineJoin: .round)
-                    context.stroke(path, with: .color($0.color), style: strokeStyle)
-                }
-            } // minimunDistance is the initial point when the line starts
-            .gesture(
-                DragGesture(minimumDistance: 0, coordinateSpace: .local)
-                    .onChanged({ value in
-                        let position = value.location
-                        // if no value has found/no line append a new start to create line
-                        if value.translation == .zero {
-                            lines.append(Line(points: [position], color: selectedColor))
-                        } else { // otherwise append to last point to create a line
-                            guard let lastIndex = lines.indices.last else { return }
-                            lines[lastIndex].points.append(position)
-                        }
-                    })
-            )
+            CanvasView(lines: $lines, selectedColor: $selectedColor)
         }
     }
